roachtest: move tests helpers to `tests/`
Release note: None
