+++ pkg/cmd/roachtest/inconsistency.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/registry"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/tests"
 	_ "github.com/lib/pq"
 )
 
@@ -48,7 +49,7 @@ func runInconsistency(ctx context.Context, t test.Test, c cluster.Cluster) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		waitForFullReplication(t, db)
+		tests.WaitFor3XReplication(t, db)
 		_, db = db.Close(), nil
 	}
