+++ pkg/cmd/roachtest/indexes.go
@@ -21,6 +21,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/registry"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/spec"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/tests"
 	"github.com/cockroachdb/cockroach/pkg/util/retry"
 )
 
@@ -72,7 +73,7 @@ func registerNIndexes(r registry.Registry, secondaryIndexes int) {
 					t.L().Printf("checking replica balance")
 					retryOpts := retry.Options{MaxBackoff: 15 * time.Second}
 					for r := retry.StartWithCtx(ctx, retryOpts); r.Next(); {
-						waitForUpdatedReplicationReport(ctx, t, conn)
+						tests.WaitForUpdatedReplicationReport(ctx, t, conn)
 
 						var ok bool
 						if err := conn.QueryRowContext(ctx, `
