+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "disk_stall.go",
         "jepsen.go",
         "predecessor_version.go",
+        "util.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachtest/tests",
     visibility = ["//visibility:public"],
@@ -18,6 +19,7 @@ go_library(
         "//pkg/cmd/roachtest/test",
         "//pkg/server/serverpb",
         "//pkg/util/httputil",
+        "//pkg/util/retry",
         "//pkg/util/timeutil",
         "//pkg/util/version",
         "@com_github_cockroachdb_errors//:errors",
