+++ pkg/cmd/roachtest/ycsb.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/registry"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/spec"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/tests"
 )
 
 func registerYCSB(r registry.Registry) {
@@ -48,7 +49,7 @@ func registerYCSB(r registry.Registry) {
 		c.Put(ctx, t.Cockroach(), "./cockroach", c.Range(1, nodes))
 		c.Put(ctx, t.DeprecatedWorkload(), "./workload", c.Node(nodes+1))
 		c.Start(ctx, c.Range(1, nodes))
-		waitForFullReplication(t, c.Conn(ctx, 1))
+		tests.WaitFor3XReplication(t, c.Conn(ctx, 1))
 
 		t.Status("running workload")
 		m := c.NewMonitor(ctx, t, c.Range(1, nodes))
